requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-environ>=0.11,<1.0
psycopg[binary]>=3.1,<4.0
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
